update: update suspense in `<Footer />` components
